+++ CHANGELOG.rst
@@ -9,3 +9,17 @@ Version 0.0.0
 =============
 
 - Release of DREAMS_MC with documentation.
+
+
+Version 0.0.1
+=============
+
+- Update on auto resizing of images while rendering.
+
+
+Version 0.0.2
+=============
+
+- CSS and JavaScript file contents are read and embedded directly into the html
+
+